Reorganise models for clarity
